feat: supabase 응답 처리 통합 : API에서 중복된 에러처리 로직 공통 함수로 분리

+++ src/api/comment.ts
@@ -1,32 +1,15 @@
 import { supabase } from '@/lib/supabaseClient';
 import { Comment, DeleteCommentParams } from '@/types/comment';
+import { withSupabaseHandler } from '@/api/utils';
 
 /**  영화 댓글 조회  */
 export const fetchComments = (movieId: number) =>
-  supabase
-    .from('Comments')
-    .select('*')
-    .eq('movie_id', movieId)
-    .then((res) => res.data);
+  supabase.from('Comments').select('*').eq('movie_id', movieId).then(withSupabaseHandler);
 
 /** 영화 댓글 추가 */
 export const addComment = (commentData: Comment) =>
-  supabase
-    .from('Comments')
-    .insert([commentData])
-    .then(({ data, error }) => {
-      if (error) throw error;
-      return data;
-    });
+  supabase.from('Comments').insert([commentData]).then(withSupabaseHandler);
 
 /**  영화 댓글 삭제  */
 export const deleteComment = ({ movieId, userId }: DeleteCommentParams) =>
-  supabase
-    .from('Comments')
-    .delete()
-    .eq('movie_id', movieId)
-    .eq('user_id', userId)
-    .then(({ data, error }) => {
-      if (error) throw error;
-      return data;
-    });
+  supabase.from('Comments').delete().eq('movie_id', movieId).eq('user_id', userId).then(withSupabaseHandler);
